+++ js/components/configuration.html
@@ -42,14 +42,14 @@
 							</td>
 							<td data-bind="text:sourceDialect"></td>
 							<td class="configureRadio">
-								<input type="radio" name="radioSource" data-bind="disable:!hasVocabulary, checkedValue:vocabularyUrl, checked:$component.sharedState.vocabularyUrl">
+								<input type="radio" name="radioSource" data-bind="disable:(!hasVocabulary || !$component.canChangePriority()), checkedValue:vocabularyUrl, checked:$component.sharedState.vocabularyUrl, click: $component.updateVocabPriority">
 								<span data-bind="text:version"></span>
 							</td>
 							<td class="configureRadio">
-								<input type="radio" name="radioEvidence" data-bind="disable:!hasEvidence, checkedValue:evidenceUrl, checked:$component.sharedState.evidenceUrl">
+								<input type="radio" name="radioEvidence" data-bind="disable:(!hasEvidence || !$component.canChangePriority()), checkedValue:evidenceUrl, checked:$component.sharedState.evidenceUrl, click: $component.updateEvidencePriority">
 							</td>
 							<td class="configureRadio">
-								<input type="radio" name="radioDensity" data-bind="disable:!hasResults, checkedValue:resultsUrl, checked:$component.sharedState.resultsUrl">
+								<input type="radio" name="radioDensity" data-bind="disable:(!hasResults || !$component.canChangePriority()), checkedValue:resultsUrl, checked:$component.sharedState.resultsUrl, click: $component.updateResultsPriority">
 							</td>
 							<td class="configureCheckbox">
 								<input type="checkbox" name="checkResults" data-bind="disable:!hasResults, checked:hasResults">
@@ -58,7 +58,9 @@
 					</tbody>
 				</table>
 			</div>
-
+			<div data-bind="if: isInProgress()" class="text-center">
+				<loading params="theme: 'light'" />
+			</div>
 			<div class="btn btn-sm btn-primary" data-bind="click: clearLocalStorageCache">Clear Configuration Cache</div>
 			<div data-bind="if: config.userAuthenticationEnabled">
 				<a href="#/roles" class="btn btn-sm btn-primary" style="width: 175px" data-bind="css: { disabled: !canReadRoles() }">Manage Permissions</a>
