+++ js/main.js
@@ -168,7 +168,24 @@ requirejs.config({
 });
 
 requirejs(['bootstrap'], function () { // bootstrap must come first
-	requirejs(['knockout', 'app', 'appConfig', 'webapi/AuthAPI', 'ohdsi.util', 'lscache', 'atlas-state', 'vocabularyprovider', 'director', 'search', 'localStorageExtender', 'jquery.ui.autocomplete.scroll', 'loading', 'user-bar', 'welcome'], function (ko, app, config, authApi, util, lscache, sharedState, vocabAPI) {
+	requirejs([
+		'knockout',
+		'app',
+		'appConfig',
+		'webapi/AuthAPI',
+		'webapi/SourceAPI',
+		'ohdsi.util',
+		'lscache',
+		'atlas-state',
+		'vocabularyprovider',
+		'director',
+		'search',
+		'localStorageExtender',
+		'jquery.ui.autocomplete.scroll',
+		'loading',
+		'user-bar',
+		'welcome'
+	], function (ko, app, config, authApi, sourceAPI, util, lscache, sharedState, vocabAPI) {
 		var pageModel = new app();
 		window.pageModel = pageModel;
 		ko.applyBindings(pageModel, document.getElementsByTagName('html')[0]);
@@ -187,7 +204,7 @@ requirejs(['bootstrap'], function () { // bootstrap must come first
 		var densityPriority = 0;
 
 		// initialize all service information asynchronously
-		var serviceCacheKey = 'ATLAS|' + config.api.url;
+		var serviceCacheKey = sourceAPI.getCacheKey();
 		cachedService = lscache.get(serviceCacheKey);
 
 		if (cachedService) {
@@ -224,117 +241,8 @@ requirejs(['bootstrap'], function () { // bootstrap must come first
 			}
 		} else {
 			config.api.sources = [];
-			var servicePromise = $.Deferred();
+			var servicePromise = sourceAPI.initSourcesConfig();//$.Deferred();
 			pageModel.initPromises.push(servicePromise);
-
-			$.ajax({
-				url: config.api.url + 'source/sources',
-				method: 'GET',
-				contentType: 'application/json',
-				success: function (sources) {
-					config.api.available = true;
-					var completedSources = 0;
-
-					$.each(sources, function (sourceIndex, source) {
-						source.hasVocabulary = false;
-						source.hasEvidence = false;
-						source.hasResults = false;
-						source.hasCDM = false;
-						source.vocabularyUrl = '';
-						source.evidenceUrl = '';
-						source.resultsUrl = '';
-						source.error = '';
-
-						source.initialized = true;
-						for (var d = 0; d < source.daimons.length; d++) {
-							var daimon = source.daimons[d];
-
-							// evaluate vocabulary daimons
-							if (daimon.daimonType == 'Vocabulary') {
-								source.hasVocabulary = true;
-								source.vocabularyUrl = config.api.url + 'vocabulary/' + source.sourceKey + '/';
-								if (daimon.priority >= vocabularyPriority) {
-									vocabularyPriority = daimon.priority;
-									sharedState.vocabularyUrl(source.vocabularyUrl);
-								}
-							}
-
-							// evaluate evidence daimons
-							if (daimon.daimonType == 'Evidence') {
-								source.hasEvidence = true;
-								source.evidenceUrl = config.api.url + 'evidence/' + source.sourceKey + '/';
-								if (daimon.priority >= evidencePriority) {
-									evidencePriority = daimon.priority;
-									sharedState.evidenceUrl(source.evidenceUrl);
-								}
-							}
-
-							// evaluate results daimons
-							if (daimon.daimonType == 'Results') {
-								source.hasResults = true;
-								source.resultsUrl = config.api.url + 'cdmresults/' + source.sourceKey + '/';
-								if (daimon.priority >= densityPriority) {
-									densityPriority = daimon.priority;
-									sharedState.resultsUrl(source.resultsUrl);
-								}
-							}
-
-							// evaluate cdm daimons
-							if (daimon.daimonType == 'CDM') {
-								source.hasCDM = true;
-							}
-						}
-
-						config.api.sources.push(source);
-
-						if (source.hasVocabulary) {
-							$.ajax({
-								url: config.api.url + 'vocabulary/' + source.sourceKey + '/info',
-								timeout: 20000,
-								method: 'GET',
-								contentType: 'application/json',
-								success: function (info) {
-									completedSources++;
-									source.version = info.version;
-									source.dialect = info.dialect;
-
-									if (completedSources == sources.length) {
-										lscache.set(serviceCacheKey, config.api, 720);
-										servicePromise.resolve();
-									}
-								},
-								error: function (err) {
-									completedSources++;
-									pageModel.initializationErrors++;
-									source.version = 'unknown';
-									source.dialect = 'unknown';
-									source.url = config.api.url + source.sourceKey + '/';
-									if (completedSources == sources.length) {
-										lscache.set(serviceCacheKey, config.api, 720);
-										servicePromise.resolve();
-									}
-								}
-							});
-						} else {
-							completedSources++;
-							source.version = 'not available'
-							if (completedSources == sources.length) {
-								servicePromise.resolve();
-							}
-						}
-					});
-				},
-				error: function (xhr, ajaxOptions, thrownError) {
-					config.api.available = false;
-					config.api.xhr = xhr;
-					config.api.thrownError = thrownError;
-
-					sharedState.appInitializationStatus('failed');
-					document.location = '#/configure';
-
-					servicePromise.resolve();
-				}
-			});
 		}
 
 		$.when.apply($, pageModel.initPromises).done(function () {
